rooms/floor_data: shift a split height as unsigned
The sign extension that a triangle's height goes through makes the
value negative, and shifting a negative int32_t left is undefined. The
shift now runs on the unsigned value, which lands on the same height.

Found with the undefined behaviour sanitizer.

+++ src/trx/game/rooms/floor_data.c
@@ -174,7 +174,7 @@ static int32_t M_DecodeSplit(int32_t height)
     if ((height & 0x10) != 0) {
         height |= 0xFFFFFFF0;
     }
-    return height << 8;
+    return (int32_t)((uint32_t)height << 8);
 }
 
 // Anything other than the player trips triggers as a heavy object.
